requirements.txt
os-sys
datetime
opencv-python
numpy
glob2
Pillow
keras-models
tkfilebrowser
cmake
dlib